pyproject.toml
[project]
name = "metadata-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
